move notes text and topic classes to notes
